requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.116.1
feedparser==6.0.11
fonttools==4.59.1
greenlet==3.2.4
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
kombu==5.5.4
llvmlite==0.43.0
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.4
numba==0.60.0
numpy==2.0.2
openai==1.100.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
scikit-learn==1.6.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
zipp==3.23.0
